DEPR: deprecate value_range (GH8481)
